+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.0.2]
+
+### Added
+* `Create useState` command
+
 ## [0.0.1]
 
-* Added `Create react component template` command
+### Added
+* `Create react component template` command
+